omiximo-email-automation/requirements.txt
# Omiximo Email Automation Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
imapclient>=2.3.1
flask>=3.0.0
flask-cors>=4.0.0
